requirements.txt
Pillow==8.4.0
tkinterdnd2==0.3.0
#pip install -r requirements.txt